[benc/a11y-metrics] Organize files into domain/ and platform/
