+++ packages/perseus-a11y-metrics/src/platform/command.ts
